src/main/java/com/aviaSoftware/ePlatnik/ePlatnikWebService/domain/Entity/PersonInfo/PersonInformation.java
package com.aviaSoftware.ePlatnik.ePlatnikWebService.domain.Entity.PersonInfo;

import java.util.Date;

public class PersonInformation {
	private String name;
	private String surName;
	private Date dateOfBirth;
	private String placeOfBirth;
	private int Old;
	private String pesel;
	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	public String getSurName() {
		return surName;
	}
	public void setSurName(String surName) {
		this.surName = surName;
	}
	public Date getDateOfBirth() {
		return dateOfBirth;
	}
	public void setDateOfBirth(Date dateOfBirth) {
		this.dateOfBirth = dateOfBirth;
	}
	public String getPlaceOfBirth() {
		return placeOfBirth;
	}
	public void setPlaceOfBirth(String placeOfBirth) {
		this.placeOfBirth = placeOfBirth;
	}
	public int getOld() {
		return Old;
	}
	public void setOld(int old) {
		Old = old;
	}
	public String getPesel() {
		return pesel;
	}
	public void setPesel(String pesel) {
		this.pesel = pesel;
	}
	
	
	
}
